test_o_parser/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.20
django-adminlte3==0.1.6
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
flower==2.0.0
gunicorn==21.2.0
h11==0.14.0
humanize==4.7.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
lxml==4.9.3
mysqlclient==2.2.0
outcome==1.2.0
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyaes==1.6.1
pyasn1==0.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
Telethon==1.29.2
tornado==6.3.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
